implementation of U26 and U27 commands -#36 #37  -implementation of U26 command for Command_Parser  -implementation of U27 command for Command_Parser

+++ Components/Command_Parser/source/forward_gcode.c
@@ -84,7 +84,7 @@ Std_Err forward_command_concurrently(SystemCommand_Settings* settings, SystemCom
     {
         char temp[15];
 
-        for(int i=PARAM_X; i<=PARAM_K; i<<=1)
+        for(uint16_t i=PARAM_X; i<=PARAM_K; i<<=1)
         {
             if(cmd->used_fields & i)
             {
